Update Kafka bootstrap server configuration and refactor publisher/consumer initialization
